Special case: allow (apparently) nested projects when testing (#241)
Sometimes I end up with a .Rproj file lying around in session temp directory, which prevents the creation of any projects in the tests. That, in turns, causes catastrophic test failure.

The presence of such a file will still cause one failure for 'test_that("a non-RStudio project is not recognized", ...)' but it doesn't seem worth it to address that.