Add use_rcpp_armadillo and use_rcpp_eigen; closes #421
usethis already has use_rcpp() to set up packages to use Rcpp.
Similarly, it would be helpful to have use_rcpp_armadillo() and
use_rcpp_eigen() to set up packages to use RcppArmadillo and RcppEigen
(@alexpghayes requested this feature in issue #421).
This commit adds both of these functions, as well as tests for them.

use_rcpp_armadillo() ensures Rcpp and RcppArmadillo are in "LinkingTo"
in the DESCRIPTION and that Rcpp is in "Imports", creates the src/
directory and an appropriate .gitignore, as use_rcpp() does,
sets up the proper src/Makevars and src/Makevars.win files (respecting
any existing content in those files), informs the user how to fix up
their NAMESPACE file, and instructs them to run devtools::document().

Similarly, use_rcpp_eigen() ensures Rcpp and RcppEigen are in
"LinkingTo" and "Imports" in the DESCRIPTION, creates the src/
directory and an appropriate .gitignore, as use_rcpp() does,
informs the user how to fix up their NAMESPACE file, and instructs
them to run devtools::document().