Remove leading dots in extension arguments to `fs::file_temp()`
With the current CRAN release of fs, passing a leading dot results in a
doubled dot.

    path_file(file_temp(ext = ".txt"))
    #> file4c8b5cc4167..txt
    path_file(file_temp(ext = "txt"))
    #> file4c8bb172736.txt

However this will be fixed by the upcoming release of fs, so that any leading
dots are ignored, so `ext = ".txt` and `ext = "txt"` will be equivalent.

    path_file(file_temp(ext = ".txt"))
    #> file4c8b24d9d42a.txt
    path_file(file_temp(ext = "txt"))
    #> file4c8b58f8244c.txt

Nonetheless I think it is best practice to supply the extension without
a dot, as I am pretty sure this was just a mistake when porting the
usethis code from using `tempfile()`.