pr_push() can now push to repo with no pr (#990) (#991)
* pr_push() now handles repo with no pr (#990)

Fixes #990

The is the  minimal fix, maybe too terse. A more expressive
solution comes in a separate commit, to allow reverting it if you
prefer the shorter solution.

The tests will also come in a separate commit because they test
the internal `pr_find()` instead of the exported `pr_push()`. You
may want to revert that upcomming commit too.

* Test pr_find() handles repo with no pr (#990)

* New is_new_pr() (more expressive version of #990)

* Simplify and record why this is needed

* Move into manual tests folder

* Rewrite as manual test

* Update NEWS bullet

* Update wordlist

Co-authored-by: Jennifer (Jenny) Bryan <jenny.f.bryan@gmail.com>