Remove `path = proj_get()` from from git helpers; prefer git_repo() over proj_get()
Inspired by working on principles.md in 8c8dbdf

Removing `path` as an arguement from internal git helpers where I think I should make implicit, hard-wired use of active project/repo

Prefer git_repo() to proj_get() because of the implicit check for git-repo-hood. The only reason to use proj_get() in the git helpers is when when we helping to establish the active project as a git repo.

Found a few places where I was neglecting to specify `repo = git_repo()` in gert calls. Although our preferred workflow will have user working in working directory that is (or is below) the active project, that is not an absolute requirement.