Unbork use_git_remote()
It is VERY important to use gert::git_remote_add() to set up a new remote, as opposed to get_remote_set_url(). The latter appears to work, but it does not fully set up the remote. In particular the fetch refspec is not configured, so operations like setting upstream tracking branch and merging fail with very mysterious errors.