Refactor create_local_thing()
Main change is to NOT assume that original working directory is same as original project. The original project could be `NULL`. This surfaced because of changes coming in withr.

While I was refactoring, I also refactored to be more in line with the recommended style in the new testthat "test fixtures" vignette.