Account for GitHub remote config "theirs"
Closes #1588

check_for_config() previously had no default handling for "theirs". It wasn't in the default ok_configs, nor was it considered a "maybe" or "bad" config. Therefore "theirs" would fall through to an internal error. In addition to direct calls to check_for_config() (not that many), this scenario affected many calls to target_repo() (lots of these).

I decided to include "theirs" in the "ok" configs, by default, since there are more functions where that's the correct handling than the other way around. In some of these settings, we really can work with "theirs". In others, "theirs" will be caught and challenged in a subsequent step based on `tr$can_push` being FALSE.

However, there are some places where we really do not support "theirs". That is why `target_repo()` needed to gain an `ok_configs` argument and pass it through to `check_for_config()`.