Rationalize nested projects (#1647)
* Rationalize nested projects

Closes #1643

We discourage nested projects, but it is possible, mostly for developers. So we should at least behave correctly in this code path.

Previously, a nested project would put the name of the parent project in DESCRIPTION, which causes a variety of downstream problems.

* NEWS bullet

* Use an undocumented option to bypass the interactive challenge

Useful in tests (replaces the mock) and also to a developer, like @lionel-, who might need to do this often enough to set the option.