Update man page for git_vaccinate()
I updated the man page for the git_vaccinate() command such that it documents **all** of the adjustments made to the `.gitignore`. I also reordered the list of files and directories that the command adds to the `.gitignore` such that it fits the order in the output returned after calling the command.

Sorry if my line breaks do not follow your standard conventions. I was a little unsure about that.