Update man page for git_vaccinate() (#1867)
* Update man page for git_vaccinate()

I updated the man page for the git_vaccinate() command such that it documents **all** of the adjustments made to the `.gitignore`. I also reordered the list of files and directories that the command adds to the `.gitignore` such that it fits the order in the output returned after calling the command.

Sorry if my line breaks do not follow your standard conventions. I was a little unsure about that.

* Revert "Update man page for git_vaccinate()"

This reverts commit 839512c212fb97ae2cf7ce312f10c5e907b8403b.

* Correction of update man page for git_vaccinate()

I corrected my updates suggested in the commit 839512c212fb97ae2cf7ce312f10c5e907b8403 according to the comments from @jennybc. Thank you!