Don't depend on dev pkgload
I took the dependency on an experimental branch of pkgload to convince myself that I solved the problem, because pkgload is indirectly also involved in the usethis situation.

There's a snapshot test for use_lifecycle() that is senstive to the pkgload version, but this won't be an issue for CRAN submission, since snapshot tests are skipped there, by default.