Decouple edition setting in capture vs expect_snapshot()
This is appears to be some sort of change in behaviour in withr 2.5.0 vs 3.0.0.

The key point is to capture the checklist with testthat 2e in force, while still being able to execute expect_snapshot(), which requires testhat 3e.

The changes around the DESCRIPTION file in the test package are separate and are just about making it look more like the DESCRIPTION file of a package who hasn't updated to testthat 3e.