Account for the absence of Quarto CLI (#2164)
* Account for the absence of Quarto CLI

* Apparently quarto_available() appeared in v1.5.0

I don't think minimum version is truly supported/enforced for a suggested package, so I don't expect this to be a definitive fix. But I think this might be as far as I can/should go here in usethis. Morally, it's the right solution.

This should be sufficient once the quarto package itself is being built for all CRAN flavors.