Lock and unlock branch as part of the release process (#2145)
* Lock and unlock branch as part of the release process

This ensures that there's no way to accidentally forget that you're in the middle of the release process and make a change to the repo.

Part of #1946

* Add to reference index

* Apply suggestions from code review

Co-authored-by: Davis Vaughan <davis@rstudio.com>

* Redocument/update snapshot tests

* Document permissions

* github_lock_branch() instead of gh_lock_branch()

Looking at other function names we use "github" (vs. "gh") everywhere except `gh_token_help()`. And that is arguably correct, since you could view that function as helping the gh package get a suitable token.

* Make github_(un)lock_branch more consistent with other functions

Also, move to different file.

* More work on docs

---------

Co-authored-by: Davis Vaughan <davis@rstudio.com>
Co-authored-by: Jenny Bryan <jenny.f.bryan@gmail.com>