Straggler commit related to #2145
I have learned the hard way that we need to more fully reverse github_lock_branch() in order to be able to delete a branch later. Learned with a PR branch. Won't come up often IRL because we will mostly lock / unlock `main`. But still better to leave things as we found them.