Add a {test} for `ui_escape_glue()` helper (#2171)
* Escape curly braces in pr_title

Added a simple helper to escape curly braces for glue (`ui_escape_glue()` in `utils-ui.R`), then applied it in the relevant `pr_*()` functions. I used Positron Assistant and Google Gemini as a test on this, then revised their fixes significantly.

Do you want to add a test for this in `manual-pr-functions.R`? Or I could automate this stuff with mocked github API calls if you want to go that far.

Fixes #2107

* Do the escape once in the sub-function toward the end.

* Add a {test} for `ui_escape_glue()` helper

And implicitly test the curly brace thing.

* Don't escape in un-glued bullets.

* Add NEWS.md bullet.

@jennybc I added the `Fixes #2107` to the description, but it didn't retroactively link to the issue, so make sure that's showing in Development to auto-close.